diagnostic-api/requirements.txt
Flask==1.1.4
Werkzeug==1.0.1
markupsafe==2.0.1
requests==2.22.0
awslambda-psycopg2==1.0.2